+++ packages/middleware-retry/src/middleware-retry.integ.spec.ts
@@ -6,7 +6,14 @@ import { requireRequestsFrom } from "../../../private/util-test/src/index";
 describe("middleware-retry", () => {
   describe(Weather.name, () => {
     it("should set retry headers", async () => {
-      const client = new Weather({ endpoint: "https://foo.bar" });
+      const client = new Weather({
+        endpoint: "https://foo.bar",
+        region: "us-west-2",
+        credentials: {
+          accessKeyId: "INTEG",
+          secretAccessKey: "INTEG",
+        },
+      });
 
       requireRequestsFrom(client).toMatch({
         hostname: "foo.bar",