+++ packages/uuid/src/randomUUID.native.ts
@@ -1,0 +1,2 @@
+// If user has provided their polyfill, like "react-native-random-uuid"
+export const randomUUID = typeof crypto !== "undefined" && crypto.randomUUID && crypto.randomUUID.bind(crypto);