+++ private/my-local-model/src/commands/GetNumbersCommand.ts
@@ -34,6 +34,8 @@ export interface GetNumbersCommandOutput extends GetNumbersResponse, __MetadataB
  * ```javascript
  * import { XYZServiceClient, GetNumbersCommand } from "xyz"; // ES Modules import
  * // const { XYZServiceClient, GetNumbersCommand } = require("xyz"); // CommonJS import
+ * // import type { XYZServiceClientConfig } from "xyz";
+ * const config = {}; // type is XYZServiceClientConfig
  * const client = new XYZServiceClient(config);
  * const input = { // GetNumbersRequest
  *   bigDecimal: Number("bigdecimal"),
@@ -53,6 +55,14 @@ export interface GetNumbersCommandOutput extends GetNumbersResponse, __MetadataB
  * @see {@link GetNumbersCommandInput} for command's `input` shape.
  * @see {@link GetNumbersCommandOutput} for command's `response` shape.
  * @see {@link XYZServiceClientResolvedConfig | config} for XYZServiceClient's `config` shape.
+ *
+ * @throws {@link CodedThrottlingError} (client fault)
+ *
+ * @throws {@link MysteryThrottlingError} (client fault)
+ *
+ * @throws {@link RetryableError} (client fault)
+ *
+ * @throws {@link HaltError} (client fault)
  *
  * @throws {@link XYZServiceServiceException}
  * <p>Base exception class for all service exceptions from XYZService service.</p>