+++ private/smithy-rpcv2-cbor-schema/src/commands/Float16Command.ts
@@ -33,6 +33,8 @@ export interface Float16CommandOutput extends Float16Output, __MetadataBearer {}
  * ```javascript
  * import { RpcV2ProtocolClient, Float16Command } from "@smithy/smithy-rpcv2-cbor-schema"; // ES Modules import
  * // const { RpcV2ProtocolClient, Float16Command } = require("@smithy/smithy-rpcv2-cbor-schema"); // CommonJS import
+ * // import type { RpcV2ProtocolClientConfig } from "@smithy/smithy-rpcv2-cbor-schema";
+ * const config = {}; // type is RpcV2ProtocolClientConfig
  * const client = new RpcV2ProtocolClient(config);
  * const input = {};
  * const command = new Float16Command(input);