+++ private/smithy-rpcv2-cbor/src/commands/SparseNullsOperationCommand.ts
@@ -34,6 +34,8 @@ export interface SparseNullsOperationCommandOutput extends SparseNullsOperationI
  * ```javascript
  * import { RpcV2ProtocolClient, SparseNullsOperationCommand } from "@smithy/smithy-rpcv2-cbor"; // ES Modules import
  * // const { RpcV2ProtocolClient, SparseNullsOperationCommand } = require("@smithy/smithy-rpcv2-cbor"); // CommonJS import
+ * // import type { RpcV2ProtocolClientConfig } from "@smithy/smithy-rpcv2-cbor";
+ * const config = {}; // type is RpcV2ProtocolClientConfig
  * const client = new RpcV2ProtocolClient(config);
  * const input = { // SparseNullsOperationInputOutput
  *   sparseStringList: [ // SparseStringList