+++ smithy-typescript-codegen/src/main/java/software/amazon/smithy/typescript/codegen/integration/HttpBindingProtocolGenerator.java
@@ -873,13 +873,14 @@ public abstract class HttpBindingProtocolGenerator implements ProtocolGenerator 
         Shape target = model.expectShape(binding.getMember().getTarget());
 
         boolean isIdempotencyToken = binding.getMember().hasTrait(IdempotencyTokenTrait.class);
-        if (isIdempotencyToken) {
-            writer
-                .addDependency(TypeScriptDependency.UUID_TYPES)
-                .addImport("v4", "generateIdempotencyToken", TypeScriptDependency.UUID);
-        }
         boolean isRequired = binding.getMember().isRequired();
-        String idempotencyComponent = (isIdempotencyToken && !isRequired) ? " ?? generateIdempotencyToken()" : "";
+
+        String idempotencyComponent = "";
+        if (isIdempotencyToken && !isRequired) {
+            writer
+                .addImport("v4", "generateIdempotencyToken", TypeScriptDependency.SMITHY_UUID);
+            idempotencyComponent = " ?? generateIdempotencyToken()";
+        }
         String memberAssertionComponent = (idempotencyComponent.isEmpty() ? "!" : "");
 
         String queryValue = getInputValue(
@@ -1004,17 +1005,12 @@ public abstract class HttpBindingProtocolGenerator implements ProtocolGenerator 
             binding.getMember(),
             target
         );
-        boolean isIdempotencyToken = binding.getMember().hasTrait(IdempotencyTokenTrait.class);
-        if (isIdempotencyToken) {
-            context.getWriter()
-                .addDependency(TypeScriptDependency.UUID_TYPES)
-                .addImport("v4", "generateIdempotencyToken", TypeScriptDependency.UUID);
-        }
 
         boolean headerAssertion = headerValue.endsWith("!");
         String headerBaseValue = (headerAssertion
             ? headerValue.substring(0, headerValue.length() - 1)
             : headerValue);
+        boolean isIdempotencyToken = binding.getMember().hasTrait(IdempotencyTokenTrait.class);
 
         if (!Objects.equals(memberLocation + "!", headerValue)) {
             String defaultValue = "";
@@ -1022,6 +1018,8 @@ public abstract class HttpBindingProtocolGenerator implements ProtocolGenerator 
                 String s = headerBuffer.get(headerKey);
                 defaultValue = " || " + s.substring(s.indexOf(": ") + 2, s.length() - 1);
             } else if (isIdempotencyToken) {
+                context.getWriter()
+                    .addImport("v4", "generateIdempotencyToken", TypeScriptDependency.SMITHY_UUID);
                 defaultValue = " ?? generateIdempotencyToken()";
             }
 
@@ -1046,6 +1044,8 @@ public abstract class HttpBindingProtocolGenerator implements ProtocolGenerator 
                 String s = headerBuffer.get(headerKey);
                 constructedHeaderValue += " || " + s.substring(s.indexOf(": ") + 2, s.length() - 1);
             } else if (isIdempotencyToken) {
+                context.getWriter()
+                    .addImport("v4", "generateIdempotencyToken", TypeScriptDependency.SMITHY_UUID);
                 constructedHeaderValue += " ?? generateIdempotencyToken()";
             } else {
                 constructedHeaderValue = headerValue;