+++ packages/service-error-classification/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## 4.2.0
+
+### Minor Changes
+
+- 45ee67f: update dist-cjs generation to use rollup
+
+### Patch Changes
+
+- Updated dependencies [45ee67f]
+  - @smithy/types@4.6.0
+
+## 4.1.2
+
+### Patch Changes
+
+- 937ac5a: make $retryable-trait errors considered transient in StandardRetryStrategyV2
+
 ## 4.1.1
 
 ### Patch Changes