+++ packages/util-defaults-mode-node/CHANGELOG.md
@@ -1,5 +1,48 @@
 # Change Log
 
+## 4.2.0
+
+### Minor Changes
+
+- 45ee67f: update dist-cjs generation to use rollup
+
+### Patch Changes
+
+- Updated dependencies [45ee67f]
+  - @smithy/config-resolver@4.3.0
+  - @smithy/credential-provider-imds@4.2.0
+  - @smithy/node-config-provider@4.3.0
+  - @smithy/property-provider@4.2.0
+  - @smithy/smithy-client@4.7.0
+  - @smithy/types@4.6.0
+
+## 4.1.5
+
+### Patch Changes
+
+- @smithy/smithy-client@4.6.5
+
+## 4.1.4
+
+### Patch Changes
+
+- @smithy/smithy-client@4.6.4
+
+## 4.1.3
+
+### Patch Changes
+
+- @smithy/smithy-client@4.6.3
+
+## 4.1.2
+
+### Patch Changes
+
+- @smithy/node-config-provider@4.2.2
+- @smithy/smithy-client@4.6.2
+- @smithy/config-resolver@4.2.2
+- @smithy/credential-provider-imds@4.1.2
+
 ## 4.1.1
 
 ### Patch Changes