+++ packages/util-endpoints/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 3.2.0
+
+### Minor Changes
+
+- 45ee67f: update dist-cjs generation to use rollup
+
+### Patch Changes
+
+- Updated dependencies [45ee67f]
+  - @smithy/node-config-provider@4.3.0
+  - @smithy/types@4.6.0
+
+## 3.1.2
+
+### Patch Changes
+
+- @smithy/node-config-provider@4.2.2
+
 ## 3.1.1
 
 ### Patch Changes