+++ packages/util-stream-browser/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 3.2.0
+
+### Minor Changes
+
+- 45ee67f: update dist-cjs generation to use rollup
+
+### Patch Changes
+
+- Updated dependencies [45ee67f]
+  - @smithy/fetch-http-handler@5.3.0
+  - @smithy/types@4.6.0
+  - @smithy/util-base64@4.2.0
+  - @smithy/util-hex-encoding@4.2.0
+  - @smithy/util-utf8@4.2.0
+
 ## 3.1.1
 
 ### Patch Changes